refactor: move input definiton to root
