+++ Assets/Scripts/Input.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 14d3e637544fd474c8cd9e07ed1a4e63
+guid: d01ceb91c4084b24f9d9e989976034f8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
